Fixed index types of Strings and Plugins for consistency.
git-svn-id: http://svn.drobilla.net/lad/slv2@315 a436a847-0d15-0410-975c-d299462d15a1

+++ hosts/lv2_jack_host.c
@@ -309,7 +309,7 @@ jack_process_cb(jack_nframes_t nframes, void* data)
 void
 list_plugins(SLV2Plugins list)
 {
-	for (size_t i=0; i < slv2_plugins_size(list); ++i) {
+	for (unsigned i=0; i < slv2_plugins_size(list); ++i) {
 		const SLV2Plugin* const p = slv2_plugins_get_at(list, i);
 		printf("%s\n", slv2_plugin_get_uri(p));
 	}
